backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
torch==2.1.1
transformers==4.36.0
gramformer==1.3
requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
numpy==1.24.3
scikit-learn==1.3.2
sentencepiece==0.1.99
accelerate==0.25.0
datasets==2.15.0
tokenizers==0.15.0
openai==1.3.7
anthropic==0.7.8
cohere==4.37
pytest==7.4.3
pytest-asyncio==0.21.1
redis==5.0.1
aioredis==2.0.1 